+++ src/axmysql/persistence/transaction.go
@@ -6,7 +6,7 @@ import (
 	"github.com/jmalloc/ax/src/ax/persistence"
 )
 
-// Tx is an implementation of persistence.Tx that wraps an SQL transaction.
+// Tx is a MySQL-backed implementation of Ax's persistence.Tx interface.
 type Tx struct {
 	ds    *DataStore
 	sqlTx *sql.Tx
@@ -15,4 +15,10 @@ type Tx struct {
 // DataStore returns the DataStore that the transaction operates on.
 func (tx *Tx) DataStore() persistence.DataStore {
 	return tx.ds
+}
+
+// ExtractTx returns the SQL transaction within tx.
+// It panics if tx is not a *Tx.
+func ExtractTx(tx persistence.Tx) *sql.Tx {
+	return tx.(*Tx).sqlTx
 }
