+++ src/axmysql/saga.go
@@ -2,12 +2,16 @@ package axmysql
 
 import (
 	"github.com/jmalloc/ax/src/ax/saga/mapping/keyset"
+	"github.com/jmalloc/ax/src/ax/saga/persistence/crud"
 	"github.com/jmalloc/ax/src/ax/saga/persistence/eventsourcing"
 	"github.com/jmalloc/ax/src/axmysql/saga"
 )
 
-// SnapshotRepository is a saga snapshot repository backed by an MySQL database.
-var SnapshotRepository eventsourcing.SnapshotRepository = saga.SnapshotRepository{}
+// SagaKeySetRepository is a key-set repository backed by an MySQL database.
+var SagaKeySetRepository keyset.Repository = saga.KeySetRepository{}
 
-// KeySetRepository is a key-set repository backed by an MySQL database.
-var KeySetRepository keyset.Repository = saga.KeySetRepository{}
+// SagaCRUDRepository is a CRUD saga repository backed by an MySQL database.
+var SagaCRUDRepository crud.Repository = saga.CRUDRepository{}
+
+// SagaSnapshotRepository is a saga snapshot repository backed by an MySQL database.
+var SagaSnapshotRepository eventsourcing.SnapshotRepository = saga.SnapshotRepository{}
