+++ src/axmysql/saga/crud.go
@@ -1,4 +1,4 @@
-package axmysql
+package saga
 
 import (
 	"context"
@@ -10,9 +10,9 @@ import (
 	mysqlpersistence "github.com/jmalloc/ax/src/axmysql/persistence"
 )
 
-// SagaRepository is an implementation of crud.Repository that stores saga
-// instances using SQL persistence.
-type SagaRepository struct{}
+// CRUDRepository is a MySQL-backed implementation of Ax's crud.Repository
+// interface.
+type CRUDRepository struct{}
 
 // LoadSagaInstance fetches a saga instance by its ID.
 //
@@ -21,7 +21,7 @@ type SagaRepository struct{}
 //
 // It panics if the repository is not able to enlist in tx because it uses a
 // different underlying storage system.
-func (r SagaRepository) LoadSagaInstance(
+func (r CRUDRepository) LoadSagaInstance(
 	ctx context.Context,
 	ptx persistence.Tx,
 	id saga.InstanceID,
@@ -69,7 +69,7 @@ func (r SagaRepository) LoadSagaInstance(
 //
 // It panics if the repository is not able to enlist in tx because it uses a
 // different underlying storage system.
-func (r SagaRepository) SaveSagaInstance(
+func (r CRUDRepository) SaveSagaInstance(
 	ctx context.Context,
 	ptx persistence.Tx,
 	i saga.Instance,
@@ -89,7 +89,7 @@ func (r SagaRepository) SaveSagaInstance(
 }
 
 // insertInstance inserts a new saga instance.
-func (SagaRepository) insertInstance(
+func (CRUDRepository) insertInstance(
 	ctx context.Context,
 	tx *sql.Tx,
 	i saga.Instance,
@@ -117,7 +117,7 @@ func (SagaRepository) insertInstance(
 
 // updateInstance updates an existing saga instance.
 // It returns an error if i.Revision is not the current revision.
-func (SagaRepository) updateInstance(
+func (CRUDRepository) updateInstance(
 	ctx context.Context,
 	tx *sql.Tx,
 	i saga.Instance,
